pydantic_framework/pyproject.toml
[project]
name = "finance-pydantic-agent"
version = "0.1.0"
description = "Finance LLM agent powered by pydantic-ai"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Finance LLM Tooling" },
]
dependencies = [
  "pydantic-ai>=0.4.0",
  "pydantic-ai-slim[openai]>=0.4.0",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "python-dotenv>=1.0",
  "yfinance>=0.2.37",
  "pandas>=2.2",
  "numpy>=1.26",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1",
  "pytest-asyncio>=0.24",
  "ruff>=0.6",
]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
line-length = 100
target-version = "py310"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
